Add files via upload
N-grams are often touted as valuable data products for data mining text. With a plethora of papers promoting the allure of n-grams, and a few that even use LLR, I built this program to quickly assess the usefulness of LLR data on bi-grams for a given corpus of text. YMMV.